thematic_analysis_app_with_graph/requirements.txt
streamlit
openai
langchain
faiss-cpu
python-docx
pdfplumber
python-dotenv
matplotlib